+++ GramsDataObj/include/MCScintHits.h
@@ -48,8 +48,8 @@ namespace grams {
     int HitID() const { return hitID; }
     int PDGCode() const { return pdgCode; }
     int Identifier() const { return volumeID; }
-    int E() const { return energy; }
-    int Energy() const { return energy; } // an extra accessor can't hurt
+    double E() const { return energy; }
+    double Energy() const { return energy; } // an extra accessor can't hurt
 
     // For users who prefer to work with ROOT's 4D vectors, and for
     // those who don't:
